requirements.txt
tf-nightly
mne
numpy
pandas
matplotlib
scikit-learn
seaborn

pyserial
scipy